modified .resize()
I think the resize event can occur more than once in a single window
resize. After a resize, the score would increase exponentially when you
killed an enemy. I think this was due to multiple binds of the click. My
theory is that the bullets were all drawn on top of one another (so it
looks like one bullet) but in fact, it was multiple, causing the drastic
score jumps. Score and resize act as expected.